Requests for change - added barcode and date
SimpleTableModel is no longer simple, but it is super generic.
You can now paste two columns into the application. The app states
explicitly that the column on the right should be the date.
The value from the left column goes onto the label as text and
as a barcode.
The value from the right column is treated as text but is expected
to be a date.
The properties should have keys for
  field_name = barcode_text
  field_barcode = barcode
  field_date = date
whose values are the names of fields in the label template.
Config now lists template_id 207, which is the cgap tiny label template
on dev pmb.